Add dynamic control over admin users
A new permission role `owners` is a fixed list of users
that can be set in the static configuration. These users
are the only ones getting access to the dynamic admin
list which is save in the state.
To make this feature usable, the direct-message handling
was removed in favor of being able to do admin & owner
commands anywhere where the bot is reachable.